Avoid Sequelize's built-in associations, remove Q
Sequelize's associations were causing major performance issues and lead
to difficult, fragile, and complex code to maintain associations
properly. This was a result of how we receive entries from the API
combined with the immaturity of Sequelize's API.

In place, we are now using thing names to form "fake" associations
between columns. This makes a lot more sense for our use-case.

In related news, rewriting these sections caused us to finish the
phasing out of `Q` and promises in favor of the `async` library.